+++ src/Storage/Caching/ChangeMonitor.cs
@@ -1,13 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Globalization;
-using System.Linq;
 using System.Runtime.Caching;
-using System.Text;
-using System.Threading.Tasks;
-using SenseNet.ContentRepository.Storage.Caching.Dependency;
 
-namespace SenseNet.ContentRepository.Storage.Caching.Builtin
+namespace SenseNet.ContentRepository.Storage.Caching
 {
     internal abstract class ChangeMonitorBase : ChangeMonitor
     {
