+++ src/Storage/Caching/PathChangeMonitor.cs
@@ -1,11 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SenseNet.ContentRepository.Storage.Caching.Dependency;
+﻿using SenseNet.ContentRepository.Storage.Caching.Dependency;
 
-namespace SenseNet.ContentRepository.Storage.Caching.Builtin
+namespace SenseNet.ContentRepository.Storage.Caching
 {
     internal class PathChangeMonitor : ChangeMonitorBase
     {
